refactor: 스키마 1차 최적화 (Wiki 참고)
자세한 내용은 Wiki 참고

+++ src/main/java/com/dsadara/realestatebatchservice/domain/RealEstateRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RealEstateRepository extends JpaRepository<RealEstate, Long> {
+public interface RealEstateRepository extends JpaRepository<RealEstate, Integer> {
 }
